数据类型、引入编码

+++ backend/ruoyi-system/src/main/java/com/ruoyi/system/domain/IcesMaterialLedger.java
@@ -38,7 +38,7 @@ public class IcesMaterialLedger extends BaseEntity {
     /**
      * 库存量
      */
-    private Long mlStock;
+    private Float mlStock;
     /**
      * 已删除
      */
